feat: Implement quickbind library, add more functions
QuickBind is a simple, header-only, C++ library that provides helper
functions to simplify the Win32 binding process. It also includes some
macros for handling std::optional values.

Additionally, this commit adds the bare minimum functions needed to
create a window. The demo package has also been updated to create an
empty window. It will be expanded in the future.